+++ common/src/main/java/de/maxhenkel/voicechat/VoicechatClient.java
@@ -15,7 +15,6 @@ import de.maxhenkel.voicechat.voice.client.MacOSUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.server.packs.repository.Pack;
-import net.minecraft.server.packs.repository.PackSource;
 
 import java.io.IOException;
 import java.nio.file.Files;
@@ -46,9 +45,9 @@ public abstract class VoicechatClient {
         BLACK_ICONS = new VoiceChatResourcePack("black_icons", new TranslatableComponent("resourcepack.voicechat.black_icons"));
 
         ClientCompatibilityManager.INSTANCE.addResourcePackSource(Minecraft.getInstance().getResourcePackRepository(), (Consumer<Pack> consumer, Pack.PackConstructor packConstructor) -> {
-            consumer.accept(Pack.create(CLASSIC_ICONS.getName(), false, () -> CLASSIC_ICONS, packConstructor, Pack.Position.TOP, PackSource.BUILT_IN));
-            consumer.accept(Pack.create(WHITE_ICONS.getName(), false, () -> WHITE_ICONS, packConstructor, Pack.Position.TOP, PackSource.BUILT_IN));
-            consumer.accept(Pack.create(BLACK_ICONS.getName(), false, () -> BLACK_ICONS, packConstructor, Pack.Position.TOP, PackSource.BUILT_IN));
+            consumer.accept(CLASSIC_ICONS.toPack());
+            consumer.accept(WHITE_ICONS.toPack());
+            consumer.accept(BLACK_ICONS.toPack());
         });
 
         if (VoicechatClient.CLIENT_CONFIG.macosMicrophoneWorkaround.get() && Platform.isMac()) {